Create new class Shop
